Move view class to its own file
